+++ Source/Web/MvcProject.Web/Areas/Administration/ViewModels/Products/ProductViewModel.cs
@@ -86,7 +86,7 @@
         [Range(ValidationConstants.ProductAllTimeAverageRatingMin, ValidationConstants.ProductAllTimeAverageRatingMax)]
         public double? AllTimeAverageRating
         {
-            get { return this.Votes.Any() ? (double)this.Votes.Average(v => v.VoteValue) : default(double?); }
+            get { return this.Votes.Any() ? this.Votes.Average(v => v.VoteValue) : default(double?); }
         }
 
         public ICollection<CommentDetailsForProductViewModel> Comments
